BinaryRegion: revised property mechanism, property values extended to all object types, keys must be string

+++ imagingbook-plugins-book/src/main/java/Ch08_Binary_Regions/Major_Axis_Demo.java
@@ -104,8 +104,8 @@ public class Major_Axis_Demo implements PlugInFilter {	// TODO: convert to overl
 				Pnt2d xc = r.getCenter();
 				double x0 = xc.getX();
 				double y0 = xc.getY();
-				double x1 = x0 + DrawingScale * r.getProperty("dx");
-				double y1 = y0 + DrawingScale * r.getProperty("dy");
+				double x1 = x0 + DrawingScale * (double) r.getProperty("dx");
+				double y1 = y0 + DrawingScale * (double) r.getProperty("dy");
 				ola.addShape(xc.getShape(0.05 * DrawingScale));
 				ola.addShape(new Line2D.Double(x0, y0, x1, y1));
 			}
